Accelerometer input method working and SettingMenu class started

+++ app/src/main/java/com/joshuawyllie/platformer/input/InputManager.java
@@ -11,7 +11,7 @@ public class InputManager {
         ACCELEROMETER
     }
 
-    public void update(float dt) {}
+    public void update(double dt) {}
     public void onStart() {}
     public void onStop() {}
     public void onPause() {}
